requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pyjwt>=2.8.0
httpx>=0.27.0
pydantic>=2.0.0
discord.py>=2.3.2
